shared/offer: move From and OfferModal enums to enums file
