+++ src/features/auth/pages/login/Login.native.test.tsx
@@ -25,7 +25,7 @@ import { EmptyResponse } from 'libs/fetch'
 import { firebaseAnalytics } from 'libs/firebase/analytics'
 import * as useFeatureFlagAPI from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { storage } from 'libs/storage'
-import { From } from 'shared/offer/components/AuthenticationModal/fromEnum'
+import { From } from 'shared/offer/enums'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { server } from 'tests/server'
 import { fireEvent, render, act, screen } from 'tests/utils'
