+++ src/features/offer/enums.ts
@@ -1,11 +1,3 @@
-export enum OfferModal {
-  APPLICATION_PROCESSING,
-  AUTHENTICATION,
-  BOOKING,
-  ERROR_APPLICATION,
-  FINISH_SUBSCRIPTION,
-}
-
 export enum ReportSteps {
   REPORT_OFFER_DESCRIPTION = 0,
   REPORT_OFFER_REASON = 1,
