+++ src/shared/offer/components/AuthenticationModal/AuthenticationModal.native.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { navigate } from '__mocks__/@react-navigation/native'
 import { analytics } from 'libs/analytics'
-import { From } from 'shared/offer/components/AuthenticationModal/fromEnum'
+import { From } from 'shared/offer/enums'
 import { fireEvent, render, waitFor } from 'tests/utils'
 
 import { AuthenticationModal } from './AuthenticationModal'
