--- src/shared/offer/components/AuthenticationModal/fromEnum.tsx
@@ -1,4 +1,0 @@
-export enum From {
-  BOOKING = 'booking',
-  FAVORITE = 'favorite',
-}
