+++ src/shared/offer/enums.ts
@@ -1,0 +1,12 @@
+export enum From {
+  BOOKING = 'booking',
+  FAVORITE = 'favorite',
+}
+
+export enum OfferModal {
+  APPLICATION_PROCESSING,
+  AUTHENTICATION,
+  BOOKING,
+  ERROR_APPLICATION,
+  FINISH_SUBSCRIPTION,
+}
